break-out GroupingLegend
